# target/devguide/hugo/static/specification/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "The Electrum Prepaid Utility Service describes an interface for facilitating\
    \ the purchase of prepaid \n\t\t\ttokens for utilities such as electricity and\
    \ water. One or more tokens for a requested value are issued by \n\t\t\tan upstream\
    \ provider (typically a payments aggregator) against a specified meter number.\
    \ The tokens (each \n\t\t\tbeing a sequence of usually 20 digits) are printed\
    \ on the customer receipt and are input manually on the \n\t\t\tmeter's keypad\
    \ for the value to be redeemed."
  version: "1.0.0"
  title: "iPrepaid Utility Service Interface"
  contact:
    name: "Electrum API"
    url: "http://io.electrum"
    email: "api@electrum.co.za"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "sandbox.electrum.io"
basePath: "/prepaidutility/v1"
schemes:
- "https"
paths:
...
  /tokenReprints/{reprintId}:
    post:
      summary: "Requests a reprint of a token"
      description: "Requests a reprint of a token that was previously issued for a\
        \ specified meter. The request can be for either the last token issued for\
        \ that meter, or for a specific transaction reference, depending on what the\
        \ provider supports."
      operationId: "requestTokenReprint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "reprintId"
        in: "path"
        description: "The randomly generated UUID of this request."
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "A token reprint request."
        required: true
        schema:
          $ref: "#/definitions/TokenReprintRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/PurchaseResponse"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/PurchaseResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/PurchaseResponse"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/PurchaseResponse"
        504:
          description: "Gateway Timeout"
          schema:
            $ref: "#/definitions/PurchaseResponse"
securityDefinitions:
  httpBasic:
    description: "All requests require HTTP basic authentication, where user name\
      \ equals the value used in\n                    MessageId.institutionId and\
      \ password has been agreed with Electrum."
    type: "basic"
definitions:
  Amount:
    type: "object"
    required:
    - "amount"
    - "currency"
    properties:
      amount:
        type: "integer"
        format: "int64"
        description: "Amount in minor denomination (e.g. cents)."
      currency:
        type: "string"
        description: "Three digit numeric ISO 4217 currency code (e.g. South African\
          \ Rand is encoded as 710)."
...
  Customer:
    type: "object"
    properties:
      firstName:
        type: "string"
        description: "The customer's first name(s)"
        minLength: 0
        maxLength: 40
      lastName:
        type: "string"
        description: "The customer's last name"
        minLength: 0
        maxLength: 40
      address:
        type: "string"
        description: "The customer's address"
        minLength: 0
        maxLength: 80
    description: "A bill payment customer"
  DebtRecoveryCharge:
    type: "object"
    required:
    - "amount"
    - "balance"
    - "description"
    - "tax"
    properties:
      amount:
        description: "Amount charged for debt recovery, in minor denomination."
        $ref: "#/definitions/Amount"
      tax:
        description: "Tax levied on this charge."
        $ref: "#/definitions/Tax"
      description:
        type: "string"
        description: "Description of this debt recovery charge (e.g. municipal rates\
          \ arrears)."
      balance:
        description: "Remaining balance on this account."
        $ref: "#/definitions/Amount"
      receiptNum:
        type: "string"
        description: "Receipt number for this charge."
...
  Meter:
    type: "object"
    required:
    - "meterId"
    properties:
      meterId:
        type: "string"
        description: "Unique identifier for the meter (e.g. serial number)."
      serviceType:
        type: "string"
        description: "Type of service dispensed by this meter (e.g. electricity, water\
          \ or gas)."
      supplyGroupCode:
        type: "string"
        description: "Code used to identify a group of suppliers."
      keyRevisionNum:
        type: "string"
        description: "Key revision number of the meter."
      tariffIndex:
        type: "string"
        description: "Code representing the tariff associated with this meter."
      tokenTechCode:
        type: "string"
        description: "The means by which tokens are supplied for this meter (typically\
          \ either numeric token or magnetic track)."
      algorithmCode:
        type: "string"
        description: "Code used to identify token encryption algorithm used by the\
          \ meter."
      newSupplyGroupCode:
        type: "string"
        description: "New supply group code. Only relevant if this has been changed\
          \ by the utility and a key change token has been issued."
      newKeyRevisionNumber:
        type: "string"
        description: "New key revision number. Only relevant if this has been changed\
          \ by the utility and a key change token has been issued."
      newTariffIndex:
        type: "string"
        description: "New tariff index. Only relevant if this has been changed by\
          \ the utility and a key change token has been issued."
...
  PurchaseResponse:
    type: "object"
    properties:
      meter:
        description: "Details of the meter."
        $ref: "#/definitions/Meter"
      customer:
        description: "Details of the person or organization to whom the meter belongs."
        $ref: "#/definitions/Customer"
      utility:
        description: "Details of the utility which manages this meter's account."
        $ref: "#/definitions/Utility"
      utilityType:
        type: "string"
        description: "Type of utility purchase being requested (e.g. electricity,\
          \ water, gas)."
      tokens:
        type: "array"
        description: "List of tokens issued for the purchase."
        items:
          $ref: "#/definitions/Token"
      debtRecoveryCharges:
        type: "array"
        description: "List of charges that have been levied in order to reclaim outstanding\
          \ debts associated with the meter."
        items:
          $ref: "#/definitions/DebtRecoveryCharge"
      serviceCharges:
        type: "array"
        description: "List of service charges levied against this meter."
        items:
          $ref: "#/definitions/ServiceCharge"
...
  ServiceCharge:
    type: "object"
    required:
    - "amount"
    - "description"
    - "tax"
    properties:
      amount:
        description: "Amount charged for the service."
        $ref: "#/definitions/Amount"
      tax:
        description: "Tax levied on this charge."
        $ref: "#/definitions/Tax"
      description:
        type: "string"
        description: "Description of this service charge (e.g. connection fee)."
  TariffBlock:
    type: "object"
    required:
    - "rate"
    - "units"
    properties:
      units:
        type: "number"
        description: "Number of units in this tariff block."
      rate:
        type: "number"
        description: "Tariff rate for this block in minor denomination."
  Tax:
    type: "object"
    required:
    - "amount"
    - "description"
    properties:
      amount:
        description: "Amount of tax levied in minor denomination."
        $ref: "#/definitions/Amount"
      description:
        type: "string"
        description: "Description of the type of tax (e.g. VAT)."
      rate:
        type: "number"
        description: "Rate at which tax is levied, expressed as a percentage."
...
  Token:
    type: "object"
    required:
    - "amount"
    - "receiptNum"
    - "tariffCalc"
    - "tax"
    - "token"
    - "tokenType"
    - "units"
    properties:
      tokenType:
        type: "string"
        description: "Type of token, namely standard (STD), basic service support\
          \ tariff (BSST), refund (REFUND), key change (KC)."
        enum:
        - "STD"
        - "BSST"
        - "REFUND"
        - "KC"
      units:
        type: "number"
        description: "Number of units redeemable by this token."
      amount:
        description: "Monetary value of the token, exclusive of tax (value set to\
          \ 0 for free token)."
        $ref: "#/definitions/Amount"
      tax:
        description: "Tax component of this token purchase (value set to 0 for free\
          \ token)."
        $ref: "#/definitions/Tax"
      receiptNum:
        type: "string"
        description: "Receipt number issued for the transaction."
      token:
        type: "string"
        description: "Numeric sequence to be entered into the meter to redeem the\
          \ token's value."
      tariffCalc:
        type: "array"
        description: "List of tariff blocks determining how the overall tariff is\
          \ calculated."
        items:
          $ref: "#/definitions/TariffBlock"
  TokenReprintRequest:
    type: "object"
    required:
    - "meterId"
    properties:
      meterId:
        type: "string"
        description: "Unique identifier (e.g. serial number) of the meter for which\
          \ token reprint is requested."
      track2Data:
        type: "string"
        description: "Track 2 data stored on the magnetic stripe of a card that is\
          \ supplied with certain meters. This data contains all meter details. It\
          \ can be used as an alternative means of input at PoS and may be required\
          \ by some provders."
      originalRef:
        type: "string"
        description: "Reference or receipt number for the original token in the case\
          \ that a specific token is requested. If this field is not present, then\
          \ the last-issued token will be requested from the provider."
  Utility:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name of the utility."
      address:
        type: "string"
        description: "Physical address of the utility."
      vatRegNum:
        type: "string"
        description: "VAT registration number of the utility."
      clientId:
        type: "string"
        description: "Identifier assigned by the utility to the client connecting\
          \ to the utility's token issuing service."
      message:
        type: "string"
        description: "Message send by the utility to be displayed on the customer\
          \ receipt."